requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.2.0
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.2.1
requests==2.28.2
sqlparse==0.4.2
stripe==5.1.1
urllib3==1.26.14
